requirements.txt
# LLM関連
ollama-python>=0.1.0
transformers>=4.36.0
torch>=2.1.0

# 音声処理
sounddevice>=0.4.6
numpy>=1.24.0
scipy>=1.11.0
librosa>=0.10.0

# アバター制御
bpy>=3.6.0
# vrm>=0.0.1  # 一時的に除外

# 配信関連
pytchat>=0.0.6
websockets>=11.0.3
requests>=2.31.0

# システム監視
psutil>=5.9.0
GPUtil>=1.4.0

# 開発ツール
pytest>=7.4.0
black>=23.7.0
mypy>=1.5.0
pydantic>=2.4.0 